+++ openretina/optimization/optimizer.py
@@ -1,41 +1,9 @@
 from typing import Callable, List, Optional
-from sys import maxsize
 
 import torch
 from torch import Tensor
 
-
-class OptimizationStopper:
-    def __init__(self, max_epochs: Optional[int] = None):
-        if max_epochs is None:
-            self.max_epochs = maxsize
-        else:
-            self.max_epochs = max_epochs
-
-    def early_stop(self, loss: float) -> bool:
-        return False
-
-
-class EarlyStopper(OptimizationStopper):
-    def __init__(self,
-                 max_epochs: Optional[int] = None,
-                 patience: int = 1,
-                 min_delta: float = 0.0):
-        super().__init__(max_epochs)
-        self._patience = patience
-        self._min_delta = min_delta
-        self._counter = 0
-        self._min_loss = float('inf')
-
-    def early_stop(self, loss: float) -> bool:
-        if loss < self._min_loss:
-            self._min_loss = loss
-            self._counter = 0
-        elif loss > (self._min_loss + self._min_delta):
-            self._counter += 1
-            if self._counter >= self._patience:
-                return True
-        return False
+from openretina.optimization.optimization_stopper import OptimizationStopper
 
 
 def optimize_stimulus(
